requirements.txt
aiofiles==0.8.0
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.6.1
async-timeout==4.0.2
bcrypt==3.2.2
blinker==1.5
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
fastapi-mail==1.1.4
flake8==5.0.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
passlib==1.7.4
Pillow==9.2.0
pyasn1==0.4.8
pycodestyle==2.9.0
pycparser==2.21
pydantic==1.9.1
pyflakes==2.5.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.20.4
typing_extensions==4.3.0
uvicorn==0.18.2
